pyproject.toml
[tool.poetry]
name = "tele_bot"
version = "0.1.0"
description = ""
authors = ["AJ <akarsh.1995.02@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7"
python-telegram-bot = "^13.4.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
